change
change

+++ database/migrations/2018_09_22_073005_create_my_classes_table.php
@@ -16,7 +16,7 @@ class CreateMyClassesTable extends Migration
         Schema::create('my_classes', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100);
-            $table->unsignedInteger('class_type_id')->nullable();
+            $table->unsignedBigInteger('class_type_id')->nullable();
             $table->timestamps();
         });
