+++ database/migrations/2018_11_03_210758_create_payments_table.php
@@ -19,7 +19,7 @@ class CreatePaymentsTable extends Migration
             $table->integer('amount');
             $table->string('ref_no', 100)->unique();
             $table->string('method', 100)->default('cash');
-            $table->unsignedInteger('my_class_id')->nullable();
+            $table->unsignedBigInteger('my_class_id')->nullable();
             $table->string('description')->nullable();
             $table->string('year');
             $table->timestamps();
