+++ database/migrations/2018_11_03_210817_create_payment_records_table.php
@@ -15,8 +15,8 @@ class CreatePaymentRecordsTable extends Migration
     {
         Schema::create('payment_records', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('payment_id');
-            $table->unsignedInteger('student_id');
+            $table->unsignedBigInteger('payment_id');
+            $table->unsignedBigInteger('student_id');
             $table->string('ref_no', 100)->unique()->nullable();
             $table->integer('amt_paid')->nullable();
             $table->integer('balance')->nullable();
